彻底去除了guava依赖了，现在为止完全不依赖guava了

+++ zframework_core/src/main/java/vo/zframework/configuration/properties/ServerConfigurationProperties.java
@@ -716,7 +716,7 @@ public class ServerConfigurationProperties {
 		this.sessionMaxActive = sessionMaxActive;
 	}
 
-	public long getSessionMaxTimeout() {
+	public int getSessionMaxTimeout() {
 		return this.sessionMaxTimeout;
 	}
